fix: initial fix

+++ ai_agents/agents/ten_packages/extension/deepgram_asr_python/extension.py
@@ -81,8 +81,8 @@ class DeepgramASRExtension(AsyncASRBaseExtension):
         if not sentence:
             return
 
-        start_ms = result.start * 1000  # convert seconds to milliseconds
-        duration_ms = result.duration * 1000  # convert seconds to milliseconds
+        start_ms = int(result.start * 1000)  # convert seconds to milliseconds
+        duration_ms = int(result.duration * 1000)  # convert seconds to milliseconds
 
         is_final = result.is_final
         self.ten_env.log_info(
